test(seo): split sitemap and section-guard tests into their own file
generate-llms-txt.test.ts is already past the 800-line file limit, so adding
the new coverage there tripped the quality gate's oversized-file regression.
Move it to generate-llms-txt-sitemap.test.ts, which also draws a cleaner
line: URL discovery and output validation on one side, the text-extraction
and formatting helpers on the other. Pure move — no assertions changed.
